requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
supabase
google-generativeai
pillow